Cody Case
Works perfectly now.  Managed to get the program to reset after it asks
the customer to exit or pull to the next spot.  Fixed an issue where the
Exit reader was deleting every instance of a customer number from the
reservation table.  Modified some of the text displaying.  I used a lot
of event handlers for Text change to get the program to reset so try not
to change the wording of the program.  Just need words centered, and the
text formatted for better visual performance.